@@ -5,20 +5,20 @@ Reporters included with `miette`.
 #[allow(unreachable_pub)]
 pub use debug::*;
 #[allow(unreachable_pub)]
-#[cfg(feature = "fancy-no-backtrace")]
+#[cfg(feature = "fancy-base")]
 pub use graphical::*;
 #[allow(unreachable_pub)]
 pub use json::*;
 #[allow(unreachable_pub)]
 pub use narratable::*;
 #[allow(unreachable_pub)]
-#[cfg(feature = "fancy-no-backtrace")]
+#[cfg(feature = "fancy-base")]
 pub use theme::*;
 
 mod debug;
-#[cfg(feature = "fancy-no-backtrace")]
+#[cfg(feature = "fancy-base")]
 mod graphical;
 mod json;
 mod narratable;
-#[cfg(feature = "fancy-no-backtrace")]
+#[cfg(feature = "fancy-base")]
 mod theme;
