@@ -3,27 +3,34 @@ use crate::{MietteError, MietteSpanContents, SourceCode, SpanContents};
 /// Utility struct for when you have a regular [`SourceCode`] type that doesn't
 /// implement `name`. For example [`String`]. Or if you want to override the
 /// `name` returned by the `SourceCode`.
-pub struct NamedSource {
-    source: Box<dyn SourceCode + 'static>,
+#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct NamedSource<S: SourceCode + 'static> {
+    source: S,
     name: String,
+    language: Option<String>,
 }
 
-impl std::fmt::Debug for NamedSource {
+impl<S: SourceCode> std::fmt::Debug for NamedSource<S> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("NamedSource")
             .field("name", &self.name)
-            .field("source", &"<redacted>");
+            .field("source", &"<redacted>")
+            .field("language", &self.language);
         Ok(())
     }
 }
 
-impl NamedSource {
+impl<S: SourceCode + 'static> NamedSource<S> {
     /// Create a new `NamedSource` using a regular [`SourceCode`] and giving
     /// its returned [`SpanContents`] a name.
-    pub fn new(name: impl AsRef<str>, source: impl SourceCode + Send + Sync + 'static) -> Self {
+    pub fn new(name: impl AsRef<str>, source: S) -> Self
+    where
+        S: Send + Sync,
+    {
         Self {
-            source: Box::new(source),
+            source,
             name: name.as_ref().to_string(),
+            language: None,
         }
     }
 
@@ -34,28 +41,38 @@ impl NamedSource {
 
     /// Returns a reference the inner [`SourceCode`] type for this
     /// `NamedSource`.
-    pub fn inner(&self) -> &(dyn SourceCode + 'static) {
-        &*self.source
+    pub fn inner(&self) -> &S {
+        &self.source
+    }
+
+    /// Sets the [`language`](SpanContents::language) for this source code.
+    pub fn with_language(mut self, language: impl Into<String>) -> Self {
+        self.language = Some(language.into());
+        self
     }
 }
 
-impl SourceCode for NamedSource {
+impl<S: SourceCode + 'static> SourceCode for NamedSource<S> {
     fn read_span<'a>(
         &'a self,
         span: &crate::SourceSpan,
         context_lines_before: usize,
         context_lines_after: usize,
     ) -> Result<Box<dyn SpanContents<'a> + 'a>, MietteError> {
-        let contents = self
-            .inner()
-            .read_span(span, context_lines_before, context_lines_after)?;
-        Ok(Box::new(MietteSpanContents::new_named(
+        let inner_contents =
+            self.inner()
+                .read_span(span, context_lines_before, context_lines_after)?;
+        let mut contents = MietteSpanContents::new_named(
             self.name.clone(),
-            contents.data(),
-            *contents.span(),
-            contents.line(),
-            contents.column(),
-            contents.line_count(),
-        )))
+            inner_contents.data(),
+            *inner_contents.span(),
+            inner_contents.line(),
+            inner_contents.column(),
+            inner_contents.line_count(),
+        );
+        if let Some(language) = &self.language {
+            contents = contents.with_language(language);
+        }
+        Ok(Box::new(contents))
     }
 }
