@@ -20,7 +20,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label("this bit here")]
             highlight: SourceSpan,
         }
@@ -52,7 +52,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -66,7 +65,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label("this bit here")]
             highlight: SourceSpan,
         }
@@ -98,7 +97,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -112,7 +110,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label("this bit here")]
             highlight: SourceSpan,
         }
@@ -144,7 +142,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -158,7 +155,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label("this bit here")]
             highlight: SourceSpan,
         }
@@ -190,7 +187,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -204,7 +200,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label]
             highlight: SourceSpan,
         }
@@ -235,7 +231,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -249,7 +244,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label("this bit here")]
             highlight: SourceSpan,
         }
@@ -281,7 +276,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -295,7 +289,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label = "x"]
             highlight1: SourceSpan,
             #[label = "y"]
@@ -347,7 +341,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -361,7 +354,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label = "these two lines"]
             highlight: SourceSpan,
         }
@@ -393,7 +386,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -407,7 +399,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label = "block 1"]
             highlight1: SourceSpan,
             #[label = "block 2"]
@@ -456,7 +448,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -472,7 +463,7 @@ mod json_report_handler {
             #[source]
             source: Inner,
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label = "block 1"]
             highlight1: SourceSpan,
             #[label]
@@ -532,7 +523,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -546,7 +536,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label = "this bit here"]
             highlight1: SourceSpan,
             #[label = "also this bit"]
@@ -588,7 +578,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -602,7 +591,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label = "this bit here"]
             highlight1: SourceSpan,
             #[label = "also this bit"]
@@ -644,7 +633,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -658,7 +646,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label = "this bit here"]
             highlight1: SourceSpan,
             #[label = "also this bit"]
@@ -700,7 +688,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -728,7 +715,6 @@ mod json_report_handler {
             "related": []
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -742,7 +728,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label("this bit here")]
             highlight: SourceSpan,
             #[related]
@@ -822,7 +808,6 @@ mod json_report_handler {
             }]
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
@@ -836,7 +821,7 @@ mod json_report_handler {
         #[diagnostic(code(oops::my::bad), help("try doing it better next time?"))]
         struct MyBad {
             #[source_code]
-            src: NamedSource,
+            src: NamedSource<String>,
             #[label("this bit here")]
             highlight: SourceSpan,
             #[related]
@@ -920,7 +905,6 @@ mod json_report_handler {
             }]
         }"#
         .lines()
-        .into_iter()
         .map(|s| s.trim_matches(|c| c == ' ' || c == '\n'))
         .collect();
         assert_eq!(expected, out);
